% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{as_POSIXct}
\alias{as_POSIXct}
\title{Convert to POSIXct type.}
\usage{
as_POSIXct(timestamps)
}
\arguments{
\item{timestamps}{Value to convert to POSIXct type. Must be in the format
%Y-%m-%d %H:%M:%S.}
}
\value{
\code{timestamps} as a POSIXct object.
}
\description{
Convert to POSIXct type.
}

