% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location.R
\name{coord_to_country}
\alias{coord_to_country}
\title{Get the ISO3 country code for a coordinate.}
\usage{
coord_to_country(lat, lon)
}
\arguments{
\item{lat}{Latitude of coordinate.}

\item{lon}{Longitude of coordinate.}
}
\value{
ISO3 country code.
}
\description{
Get the ISO3 country code for a coordinate.
}

