% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location.R
\name{coordinate_to_state}
\alias{coordinate_to_state}
\title{Get the fulll state for a coordinate.}
\usage{
coordinate_to_state(lat, lon)
}
\arguments{
\item{lat}{Latitude of coordinate.}

\item{lon}{Longitude of coordinate.}
}
\value{
Lowercase state name or \code{NA}.
}
\description{
Return value will be NA if coordinate is located outside mainland United
States. Thus Hawaii and Alaska coordinates will return NA.
}

