% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{data_frame}
\alias{data_frame}
\title{Create a data frame.}
\usage{
data_frame(col_names, n_rows)
}
\arguments{
\item{col_names}{The column names for the data frame.}

\item{n_rows}{The number of rows to be in the data frame.}
}
\value{
A data frame.
}
\description{
Create a data frame.
}

