% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date.R
\name{days_between_inclusive}
\alias{days_between_inclusive}
\title{Calculate the number of days between two days, inclusive.}
\usage{
days_between_inclusive(start_date, end_date)
}
\arguments{
\item{start_date}{A start date string.}

\item{end_date}{An end date string.}
}
\value{
The number of days between inclusive.
}
\description{
Calculate the number of days between two days, inclusive.
}

