% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location.R
\name{distance_between}
\alias{distance_between}
\title{Calculate coordinate distance.}
\usage{
distance_between(lat1, lon1, lat2, lon2)
}
\arguments{
\item{lat1}{Latitude of first coordinate.}

\item{lon1}{Longitude of the first coordinate.}

\item{lat2}{Latitude of second coordinate.}

\item{lon2}{Longitude of the second coordinate.}
}
\value{
The distance between the coordinates in meters
}
\description{
Calculate coordinate distance.
}

