% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date.R
\name{generate_date_block_indices}
\alias{generate_date_block_indices}
\title{Generate date block indices.}
\usage{
generate_date_block_indices(total_number_of_days, days_per_request = 31)
}
\arguments{
\item{total_number_of_days}{How many days are going to be requested.}

\item{days_per_request}{The number of days per request. Defaults to max of
31.}
}
\value{
A collection of indices where \code{result[i] + 1} represents the
start date index and \code{result[i]} represents the end date index.
}
\description{
Due to SolgarGIS restricting the size of any one request to 31 days, we must
create a collection of indices such that a proper fetch range is created
from \code{result[i] + 1} to \code{result[i]}.
}

