% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{read_csv}
\alias{read_csv}
\title{Read a SolarGIS CSV file.}
\usage{
read_csv(file, sorted = TRUE)
}
\arguments{
\item{file}{The path to a SolarGIS CSV file.}

\item{sorted}{Should the returned data frame be sorted. default value is
\code{TRUE}.}
}
\value{
A data frame with correctly formatted row names.
}
\description{
Read a SolarGIS CSV file.
}

