% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{remove_nas}
\alias{remove_nas}
\title{Remove \code{NA}s from array.}
\usage{
remove_nas(x)
}
\arguments{
\item{x}{Array to remove \code{NA}s from.}
}
\value{
An array with \code{NA}s.
}
\description{
Remove \code{NA}s from array.
}

