% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{request}
\alias{request}
\title{Request SolarGIS data.}
\usage{
request(solargis_dir, lat, lon, year, author, min_dist = 10000)
}
\arguments{
\item{solargis_dir}{The root directory where solar-gis data is stored.}

\item{lat}{The location's latitude.}

\item{lon}{The location's longitude.}

\item{year}{The year of data to be requested.}

\item{min_dist}{The minimum distance in meters from (lat, lon) that is
acceptable for use. Defaults to 10000 meters which is about 6.2 miles.}
}
\value{
A complete file path for the SolarGIS data requested.
}
\description{
The parameters of this function may be used to perform requests for multiple
sights for multiple years.
}
\author{
author The author of the request. May be an individual's name or the
project name requesting data.
}

