% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{request_remote}
\alias{request_remote}
\title{Request data from SolarGIS servers.}
\usage{
request_remote(lat, lon, start_date, end_date, api_key)
}
\arguments{
\item{lat}{The location's latitude.}

\item{lon}{The location's longitude.}

\item{start_date}{The start date for the data to be requested.}

\item{end_date}{The end date for the data to be reqeusted.}

\item{api_key}{A SolarGIS API key.}
}
\value{
A complete file path for the SolarGIS data requested.
}
\description{
Submit a single request for a single location to SolarGIS's REST API.
IMPORTANT: NO CHECKS ARE MADE WHEN THIS FUNCTION IS CALLED AGAINST PREVIOUS
REQUESTS SO THE TIME PERIOD WILL COUNT AGAINST OUR DATA UNITS.
}

