% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{rest_datadelivery_url}
\alias{rest_datadelivery_url}
\title{Compose a SolarGIS REST Data Delivery URL.}
\usage{
rest_datadelivery_url(route, api_key)
}
\arguments{
\item{route}{The route name. Will be prefixed with
\code{https://solargis.info/ws/rest/datadelivery/}.}

\item{api_key}{A SolarGIS API key.}
}
\value{
A SolarGIS url as a string.
}
\description{
Compose a SolarGIS REST Data Delivery URL.
}

