% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date.R
\name{subtract_date_ranges}
\alias{subtract_date_ranges}
\title{Subtract date ranges.}
\usage{
subtract_date_ranges(start_date, end_date, lower_date, upper_date)
}
\arguments{
\item{start_date}{The start date for requested data.}

\item{lower_date}{The lower date for known SolarGIS data.}

\item{upper_date}{The uppper date for known SolarGIS data.}

\item{start_date}{The end date for requested data.}
}
\value{
A collection of date ranges that are not included in the range
\code{lower_date} and \code{upper_date}.
}
\description{
The set difference \code{[start_date, end_date] - [lower_date, upper_date]}
where the \code{[lower_date, upper_date]} range represents dates for which
SolarGIS data is known.
}

