% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usage.R
\name{usage_calls_remaining}
\alias{usage_calls_remaining}
\title{Get the number of remaining SolarGIS API calls.}
\usage{
usage_calls_remaining(api_key)
}
\arguments{
\item{api_key}{A SolarGIS API key.}
}
\value{
The number of calls remaining or NA.
}
\description{
Get the number of remaining SolarGIS API calls.
}

