% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usage.R
\name{usage_units}
\alias{usage_units}
\title{Get the number of used SolarGIS units.}
\usage{
usage_units(api_key, start_date, end_date)
}
\arguments{
\item{api_key}{A SolarGIS API key.}

\item{start_date}{Query date range start. Format: yyyy-mm-dd.}

\item{end_date}{Query date range end. Format: yyyy-mm-dd.}
}
\value{
The number of used data units.
}
\description{
A single unit is one day of data.
}

