% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usage.R
\name{usage_units_remaining}
\alias{usage_units_remaining}
\title{Get the number of remaining SolarGIS units.}
\usage{
usage_units_remaining(api_key)
}
\arguments{
\item{api_key}{A SolarGIS API key.}

\item{start_date}{Query date range start. Format: yyyy-mm-dd.}

\item{end_date}{Query date range end. Format: yyyy-mm-dd.}
}
\value{
The number of remaining units.
}
\description{
A single unit is one day of data.
}

