% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writers.R
\name{write_meta}
\alias{write_meta}
\title{Write meta data.}
\usage{
write_meta(meta, meta_file, should_append = TRUE,
  should_use_column_names = FALSE)
}
\arguments{
\item{meta}{Meta data frame.}

\item{meta_file}{Meta CSV file.}

\item{should_append}{Should the entry be appended to existing file.}

\item{should_use_column_names}{Should column names be used.}
}
\description{
Writes meta data frame to meta file.
}

