% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writers.R
\name{write_site_data}
\alias{write_site_data}
\title{Write site data.}
\usage{
write_site_data(data, site_data_file)
}
\arguments{
\item{data}{Site data frame.}

\item{site_data_file}{Site data file.}
}
\description{
Appends data to site data file or creates a new file.
}

