% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{request}
\alias{request}
\title{Request SolarGIS data.}
\usage{
request(solargis_dir, site_id, lat, lon, start_date, end_date, author, api_key,
  min_dist = 1000)
}
\arguments{
\item{solargis_dir}{The root directory where solar-gis data is stored.}

\item{site_id}{A site identifier. Does not be unique. Unique identifier is
created with \code{site_id} and \code{location_hash} in the meta data table.}

\item{lat}{The location's latitude.}

\item{lon}{The location's longitude.}

\item{start_date}{The start date for the data to be requested.}

\item{end_date}{The end date for the data to be reqeusted.}

\item{api_key}{A SolarGIS API key.}

\item{min_dist}{The minimum distance in meters from (lat, lon) that is
acceptable for use. Defaults to 1000 meters which is about 0.62 miles.}
}
\value{
A complete file path for the SolarGIS data requested.
}
\description{
The parameters of this function may be used to perform requests for multiple
sights for multiple years.
}
\author{
author The author of the request. May be an individual's name or the
project name requesting data.
}

