% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writers.R
\name{write_meta}
\alias{write_meta}
\title{Write meta data.}
\usage{
write_meta(meta, meta_file)
}
\arguments{
\item{meta}{Meta data frame.}

\item{meta_file}{Meta CSV file.}
}
\description{
Writes meta data frame to meta file.
}

